made a config.ts where we can get app with the config to our files
